Add files via upload
Removed the Case statement and went back to basics. This code just drives the robot forward for 2 seconds, pauses for 4 seconds (so user can read the output screen) and displays the number of clicks for the left and right encoders. It will execute 5 times before stopping entirely. NOTE: I don't know what channels the 2nd encoder was hooked up to, so need to check the values where both rightEncoder and leftEncoder are instantiated.